+++ lib/src/internal/background_activation_completer.dart
@@ -1,0 +1,5 @@
+import 'base_activation_completer.dart';
+
+class AppMetricaBackgroundActivationCompleter extends BaseActivationCompleter {
+  AppMetricaBackgroundActivationCompleter(super.config);
+}